+++ song-service/src/main/kotlin/com/epam/repository/SongRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SongRepository : CrudRepository<Song, Int>
+interface SongRepository : CrudRepository<Song, Long>
